requirements.txt
blabel
weasyprint
gobject
cairocffi
pyyaml